requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.19.1
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
pillow==11.3.0
requests==2.32.5
six==1.17.0
sympy==1.14.0
torch==2.8.0
torch-vision==0.1.6.dev0
torchtext==0.18.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0

# New dependencies for ConvNeXt-RADAM
timm>=0.9.0
scikit-learn>=1.3.0

# OpenCLIP for ViT-G/14 with LAION-2B pretraining
open_clip_torch>=2.20.0
matplotlib>=3.5.0
